[Update] Modify UI and project structure
- Redesign send button
- Modify project structure
- Modify string files
- Modify prompts
