+++ app/src/main/java/levi/lin/gemini/android/utility/ModelNameListener.kt
@@ -1,4 +1,4 @@
-package levi.lin.gemini.android
+package levi.lin.gemini.android.utility
 
 interface ModelNameListener {
     fun onModelNameChanged(targetModelName: String)
