+++ app/src/main/res/values-zh-rHK/strings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Gemini</string>
-    <string name="action_go">遞交</string>
     <string name="gemini_input_label">愛你嘅魔法小幫手</string>
-    <string name="gemini_input_hint">請輸入任何文字，小幫手會友善咁回應你</string>
+    <string name="gemini_input_hint">請輸入任何文字，小幫手會友善咁回應</string>
 </resources>
