+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Gemini</string>
-    <string name="action_go">送出</string>
     <string name="gemini_input_label">愛你的魔法小幫手</string>
-    <string name="gemini_input_hint">請輸入任何文字，小幫手會親切的回應你唷</string>
+    <string name="gemini_input_hint">請輸入任何文字，小幫手會親切的回應</string>
 </resources>
