+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,5 @@
 <resources>
     <string name="app_name">Gemini</string>
-    <string name="action_go">GO</string>
-    <string name="gemini_input_label">Your lovely magic assistant</string>
-    <string name="gemini_input_hint">Please ask anything you want to know, Gemini will respond to you kindly</string>
+    <string name="gemini_input_label">Your lovely assistant</string>
+    <string name="gemini_input_hint">Please ask anything you want to know, Gemini will respond to you kindly.</string>
 </resources>
